refactor(node): extract cli parser module from main (#2602)
